requirements.txt
Django==2.2.14
django-countries==5.5
django-debug-toolbar==3.2
djangorestframework==3.12.2
django-rest-swagger==2.2.0
django-extensions==2.2.5
Pillow==8.1.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
gunicorn==19.6.0
Unipath==1.0
python-decouple==3
Markdown==2.6.6
bleach==1.4.3
psycopg2-binary==2.8.3
whitenoise==3.2